+++ src/managers/mod.rs
@@ -6,6 +6,6 @@ pub mod yay;
 
 pub use apt::check_apt;
 pub use cargo::check_cargo;
-pub use go::check_go;
+// pub use go::check_go;
 pub use snap::check_snap;
 pub use yay::check_yay;